Browser Panel & MonthView working together successfully
