+++ src/main/resources/view/DarkTheme.css
@@ -402,3 +402,16 @@
     -fx-text-fill: #000000;
     -fx-background-color: #FFFF00;
 }
+
+/* @@author guekling */
+
+#calendarPane {
+    -fx-background-color: #FFFFFF;
+}
+
+#calendarTitle {
+    -fx-font-size: 12pt;
+    -fx-font-family: "Segoe UI Semibold";
+    -fx-font-weight: bold;
+}
+
